Refactored GetOemEncoding into the XPlatAdapters
