+++ source/SilentProcessRunner/IXPlatAdapter.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Net;
+using System.Text;
 
 namespace Octopus.SilentProcessRunner
 {
@@ -9,5 +10,6 @@ namespace Octopus.SilentProcessRunner
     {
         void RunAsDifferentUser(ProcessStartInfo startInfo, NetworkCredential runAs, IDictionary<string, string>? customEnvironmentVariables);
         void TryKillProcessAndChildrenRecursively(Process process);
+        Encoding GetOemEncoding();
     }
 }
